+++ inst/include/rxode2parse_control.h
@@ -130,6 +130,7 @@
 #define Rxc_chunkSize 126
 #define Rxc_parallel 127
 #define Rxc__zeros 128
+#define Rxc_zeroVarParamHandle 129
 #define RxMv_params 0
 #define RxMv_lhs 1
 #define RxMv_state 2